ASM updates (#816)
* Revise ASM SupplyControl Contract construction

The `operator` of the SupplyControl is the address that is authorized to
mint and burn Auton supply, and is intended to be the Stabilization
Contract.

With this change, the SupplyControl contract takes an additional
argument for the `operator`.

Previously, the SupplyControl Contract was not passed the address of the
`operator` at construction time, because it was assumed that the address
of the latter would not be known at that time in the deployment
sequence. So it was necessary to set the operator of the SupplyControl
in a subsequent call to the contract to avoid circular dependency.

However, as the ASM is deployed at genesis, and the addresses of all the
protocol contracts is known at that time, there is no loneger any need
for the two-step process.

* Improve ASM terminology regarding target price

The term 'target price' is clearer and less ambiguous than the
'redemption price'.

* Make Autonity contract update ACU

If the oracle has some new prices available, the
autonity contract will call the update function in the ACU contract.

* Update solc version and enable optimization

This was necessary to resolve a stack too deep issue
being raised. The added --optimize-runs flag is set
to optimize for runtime gas cost over deployment cost.

* Supply Control contract updates

Make it following autonity protocol contracts conventions.
The operator contract is renamed to stabilizer.
The admin contract is renamed to operator.
The autonity account is passed as an argument to retrieve the operator.
Minor other naming changes.

* Add interfaces for the ASM contract

That speeds up considerably compilation by avoiding contract
import.

* Rewiring to call the autonity contract's default.

Also refactoring of the config Autonity struct because of
compiler limitations.

* Specify EVM compilation target to London

And L1 updates for deployment.

* Tidy up ASM contracts after earlier changes

This is mostly reformatting in accordance with the Solidity Style Guide.

There are small functional changes in the ACU Contract to resurrect the
view functions, and similary for the Stabilization Contract `accounts()`
function that is a convenience for liquidators.

* IAutonity.sol: Add Solidity version pragma

Match existing `Autonity.sol`.

* Activate the ACU Contract update in Autonity finalize()

* Use non-payable for Autonity in ASM constructor parameters

The ASM only uses Autonity Contract address for access control, there is
no functionality currently for send or transfer. So we don't need the
payable type.

* Pass operator and oracle to ASM contract constructors

In a recent commit, the ASM was modified so that the ACU and
Stabilization contracts only receive the address of the Autonity
Contract, and they don't receive the address of the Oracle or the
Operator directly. Instead, if they need to access the oracle or
operator address, they call `autonity.getOracle()` or
`autonity.getOperator()`.

Unfortunately, it was found that this tight coupling between the
Autonity Contract and the dependent ASM contracts makes the ASM more
difficult to test, even with mock implementations of `IAutonity`.

The main issue is the complexity of setting up the Oracle Contract in a
specific state, and then being able to call `finalize()` on it. This
will only become more tricky in future as we will need to do the same
for the ACU as well.

Instead of the indirection, this change passes the Governance Operator
address and the Oracle Contract address directly to the ASM
constructors. The downside is that we have to explicitly update these
contracts when there is a change to the operator or oracle. But at this
time it seems to be a smaller price to pay in terms of development and
maintenance effort.

* Autonity Contract updates ASM operators when changed

* Update ASM tests for Autonity integration changes

* Update generated contract bindings

* Update AutonityTest.sol for new config structure

* Fix truffle tests

---------

Co-authored-by: Youssef Azzaoui <ya@clearmatics.com>