Change the Oracle Contract symbol format
Using slash character in the symbol creates problems for users, because
the symbol cannot be used directly in URLs or UNIX filenames, which
don't permit slashes without additional quoting.

This commit changes the divider to a hyphen symbol (45 in ASCII), which
is another conventional divider that is commonly used.

This is better than no divider, because it has the advantage of being
unambiguous. If there was no divider, it is possible to have a situation
where you cannot identify the two assets in a pair without relying on
additional information that is outside the symbol.