Use a fixed version of the Ape Framework
Currently the Ape Framework version is not fixed, so developers may be
using different versions, and the CI will use the latest released
version.

This would be fine if the framework was stable. But Ape does not yet
seem to have a stable API, as there seem to be frequent breaking changes
that are not announced or documented.

To provide stability for our contract testing this change fixes Ape to a
specific version of the framework. We can then upgrade the framework
version in a deliberate manner after thorough testing and understanding
the Ape framework changes.