Generate NatSpec data together with contract ABIs
Currently, the protocol contract documentation is maintained manually in
the docs.autonity.org repository.

As part of the documentation effort for mainnet launch, we are planning
to migrate protocol contract documentation to the source code and
automatically generate documentation from NatSpec data.

Using the existing Autonity build process for generating NatSpec files
would reduce the maintenance burden on the documentation generator,
compared to maintaining a similar Makefile in the docs repository.
With the `--userdoc` and `--devdoc` Solc compiler options the NatSpec
JSONs will be available next to the ABI files with `.docdev` and
`.docuser` file extensions that can be used as inputs to the generator.